Changing to use the sample that is in samples/embeded as a unit test.
