+++ samples/embed/test.cs
@@ -5,7 +5,10 @@ class MonoEmbed {
 	[MethodImplAttribute(MethodImplOptions.InternalCall)]
 	extern static string gimme();
 
-	static void Main() {
-		Console.WriteLine (gimme ());
+	static int Main() {
+		System.Console.WriteLine(gimme ());
+		if (gimme ().Equals("All your monos are belong to us!"))
+            return 0;
+        return 100;
 	}
 }
